Refactor Routes and scenes
